taskmanager/requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
black==26.3.1
celery==5.6.3
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==6.0.4
djangorestframework==3.17.1
djangorestframework-api-key==3.1.0
flower==2.0.1
humanize==4.15.0
iniconfig==2.3.0
isort==8.0.1
kombu==5.6.2
mypy_extensions==1.1.0
mysqlclient==2.2.8
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.6
pluggy==1.6.0
prometheus_client==0.25.0
prompt_toolkit==3.0.52
Pygments==2.20.0
pytest==9.0.3
pytest-django==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytokens==0.4.1
pytz==2026.1.post1
six==1.17.0
sqlparse==0.5.5
tornado==6.5.5
tzdata==2026.1
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
